IMAGE_reccomendation/backend-rec/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
Pillow==10.0.0
requests==2.31.0
